Rename forward() to get_latent()
