+++ src/garage/np/embeddings/encoder.py
@@ -6,14 +6,26 @@ class Encoder(abc.ABC):
     """Base class of context encoders for training meta-RL algorithms."""
 
     @abc.abstractmethod
-    def forward(self, input_value):
+    def get_latent(self, input_value):
         """Encode an input value.
 
         Args:
-            input_value (numpy.ndarray): Input values of (N, input_dim) shape.
+            input_value (numpy.ndarray): Input value of (input_dim, ) shape.
 
         Returns:
             numpy.ndarray: Encoded embedding.
+
+        """
+
+    @abc.abstractmethod
+    def get_latents(self, input_values):
+        """Encode a batch of input values.
+
+        Args:
+            input_values (numpy.ndarray): Input values of (N, input_dim) shape.
+
+        Returns:
+            numpy.ndarray: Encoded embeddings.
 
         """
